unchained/requirements.txt
addict==2.2.0
atomicwrites==1.2.1
attrs==18.2.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-rest-framework-redocs==0.0.3
djangorestframework==3.8.2
drf-autodocs==0.4.4
drfdocs==0.0.11
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
graphviz==0.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
more-itertools==4.3.0
numpy==1.15.0
oauthlib==2.1.0
Pillow==5.2.0
pluggy==0.7.1
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.8.0
pytest-django==3.4.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytorchviz==0.0.1
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
torch==0.4.1
torchvision==0.2.1
uritemplate==3.0.0
urllib3==1.23